% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_files_all.R
\name{get_sample_files_all}
\alias{get_sample_files_all}
\title{Get the object containing all sample files}
\usage{
get_sample_files_all()
}
\value{
All sample files bound by row \code{sample_files_all}. See \code{?sample_files}.

\code{get_sample_files_all2()} is different in two ways: (1) uses a \code{fy_year} name,
instead of \code{fy.year}, and (2) uses the same names as in \code{sample_file_1314}.
}
\description{
Get the object containing all sample files
}
