% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_by_postcode_sex_1314.R
\docType{data}
\name{income_by_postcode_sex_1314}
\alias{income_by_postcode_sex_1314}
\title{Income by postcode and sex 2013-14}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5142 rows and 6 columns.}
\source{
\url{http://data.gov.au/dataset/ad-hoc-data-requests/resource/35f668a0-b24c-4b23-ae42-1f42eeefbaa2}
}
\usage{
income_by_postcode_sex_1314
}
\description{
On the tin.

A table of 5142 rows and 5 columns:
\describe{
\item{State}{The state or territory of the taxpayer.}
\item{Postcode}{The postcode of the taxpayer. Some small postcodes are marked Other State}
\item{Number of individuals}{The number of taxpayer-lodgers}
\item{Average taxable income}{The average taxable income filed.}
\item{Sex}{The sex as marked.}
}
}
\keyword{datasets}

