% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_by_taxable_status_residency_taxable_income_range.R
\docType{data}
\name{indiv_by_taxable_status_residency_taxable_income_range}
\alias{indiv_by_taxable_status_residency_taxable_income_range}
\title{Selected items by taxable status, residence status, and taxable income ranges}
\format{A data frame of 7236 observations and 6 columns.
\describe{
\item{Taxable_status}{Taxable or Non-Taxable.}
\item{Residency_status}{Resident or Non-resident.}
\item{Taxable_income_ranges}{Ranges of taxable income (not necessarily increasing).}
\item{variable}{A variable relating to the id columns \code{Taxable_status}, \code{Residency_status}, \code{Taxable_income_ranges}.}
\item{value}{The value of that variable for the id columns.}
\item{unit}{The units of value.}
}}
\usage{
indiv_by_taxable_status_residency_taxable_income_range
}
\description{
Various aggregated statistics of personal income tax in 2013-14.
}
\keyword{datasets}

