% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals_table3_200910.R
\docType{data}
\name{individuals_table3_200910}
\alias{individuals_table3_200910}
\title{Selected items by postcode, 2009-10}
\format{A data frame of 5 columns and 102648 observations.
\describe{
\item{Postcode}{2009-10 postcode of taxpayer.}
\item{variable}{The variable relating to value.}
\item{value}{The value of \code{variable}.}
\item{unit}{Unit: number of nonzero observations, or value in dollars.}
\item{State}{The state of the postcode.}
}}
\usage{
individuals_table3_200910
}
\description{
Selected variables of personal income tax, by postcode, and unit.
}
\keyword{datasets}

