% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_super_contr_base.R
\docType{data}
\name{max_super_contr_base}
\alias{max_super_contr_base}
\title{Maximum quarterly superannuation contribution base, by year}
\format{A data frame of 25 observations and 2 columns
\describe{
\item{fy_year}{The applicable financial year.}
\item{max_sg_per_qtr}{Maximum salary for which the superannuation guarantee is payable.}
}}
\usage{
max_super_contr_base
}
\description{
The amount of salary income above which no super guarantee payment is compulsory.
}
\keyword{datasets}
