% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_taxable_income_by_postcode_200607.R
\docType{data}
\name{mean_taxable_income_by_postcode_200607}
\alias{mean_taxable_income_by_postcode_200607}
\title{Mean taxable income by postcode, 2006-07}
\format{A data table with 1455 rows and 5 columns.}
\source{
Taxstats 2006-07 concatenated XLS files
}
\usage{
mean_taxable_income_by_postcode_200607
}
\description{
A table of 2477 rows and 2 columns:
\describe{
\item{Postcode}{A character representing the 2006-07 Postcode.}
\item{mean_taxable_income}{The mean taxable income for that postcode in 2006-07.}
}
}
\keyword{datasets}
