% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperannuationVariables.R
\name{mutate_super_vars}
\alias{mutate_super_vars}
\title{Superannuation variables}
\usage{
mutate_super_vars(.sample.file)
}
\arguments{
\item{.sample.file}{A sample file onto which new columns are to be added.}
}
\value{
A data.table
}
\description{
A way to add superannuation variables to sample files.
}

