% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupation_decoder.R
\docType{data}
\name{occupation_decoder}
\alias{occupation_decoder}
\title{Occupation decoder for sample files}
\format{A data frame of 10 observations and two columns
\describe{
\item{Occ_code}{The occupation code, key of sample files.}
\item{Occupation_description}{The English description corresponding to \code{Occ_code}.}
}}
\usage{
occupation_decoder
}
\description{
Decodes the occupation codes in the sample file
}
\keyword{datasets}

