% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{A function for simple projections of sample files}
\usage{
project(sample_file, h = 0L, fy.year.of.sample.file = "2012-13",
  WEIGHT = 50L)
}
\arguments{
\item{sample_file}{A sample file, most likely the 2012-13 sample file. It is intended that to be the most recent.}

\item{h}{An integer. How many years should the sample file be projected?}

\item{fy.year.of.sample.file}{The financial year of \code{sample_file}.}

\item{WEIGHT}{The sample weight for the sample file. (So a 2\%\ file has \code{WEIGHT} = 50.)}
}
\value{
A sample file of the same number of rows as \code{sample_file} with inflated values (including WEIGHT).
}
\description{
A function for simple projections of sample files
}

