% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_decoder.R
\docType{data}
\name{region_decoder}
\alias{region_decoder}
\title{Region decoder for sample files}
\format{A data frame of 35 observations and 2 columns.
\describe{
\item{Region}{An integer corresponding to the Region code in the sample files.}
\item{Region_description}{The region of Australia corresponding to Region code.}
}}
\usage{
region_decoder
}
\description{
Region description for each region code in the sample files.
}
\keyword{datasets}

