% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample_file_1415_synth.R
\docType{data}
\name{sample_file_1415_synth}
\alias{sample_file_1415_synth}
\title{Fictional 2014-15 sample file}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 263339 rows and 66 columns.}
\usage{
sample_file_1415_synth
}
\description{
Since the 2014-15 sample file is not public, this file has been used to test basic compatibility with it. It has the same number of rows, but that's about it. There is no guarantee that file is representative; indeed the values may be nonsensical. Contact the ATO to obtain a sample file that is useable research purposes.
}
\keyword{datasets}
