% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_file_variable_names.R
\docType{data}
\name{sample_file_variable_names}
\alias{sample_file_variable_names}
\title{Sample file variable names}
\format{A data frame of 69 observations and 3 columns.
\describe{
\item{No.}{The column number.}
\item{File column name}{The column name in the sample file.}
\item{Description}{A description of the file column name.}
\item{Tax return label}{The reference to the individuals tax return, 2013-14.}
\item{Data type}{The storage mode of the data.}
\item{Values}{The possible values.}
}}
\usage{
sample_file_variable_names
}
\description{
Full descriptions of hte variables in the sample files.
}
\keyword{datasets}

