% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_files_all.R
\name{get_sample_files_all}
\alias{get_sample_files_all}
\title{Get the object containing all sample files}
\usage{
get_sample_files_all(assign.env = parent.frame(2))
}
\value{
A promise to \code{sample_files_all}. See \code{?sample_files}.
}
\description{
Get the object containing all sample files
}

